Convert json-view to DataTree component
