+++ src/components/DataTree.module.scss
@@ -1,7 +1,3 @@
-/**
- * Based on https://github.com/pgrabovets/json-view/blob/master/src/jsonview.scss
- */
-
 .jsonContainer {
   font-family: 'Open Sans';
   font-size: 16px;
